+++ src/cljs/athens/parse_renderer.cljs
@@ -294,7 +294,7 @@
                                   (js/console.warn "Unexpected KaTeX error" e)
                                   (aset el "innerHTML" text)))))}])
      :newline (fn [_]
-                " ")}
+                [:br])}
    tree))