+++ src/cljs/athens/views/breadcrumbs.cljs
@@ -37,7 +37,12 @@
    :transition "all 0.3s ease"
    ::stylefy/manual [[:a {:text-decoration "none"
                           :cursor "pointer"
+                          :position "relative"
                           :color "inherit"}]
+                     [:* {:display "inline"
+                          :margin 0
+                          :padding 0
+                          :font-size "inherit"}]
                      [:&:last-child {:color (color :body-text-color)}]
                      [:&:hover {:flex-shrink "0"
                                 :color (color :link-color)}]