+++ src/cljs/athens/views/pages/node_page.cljs
@@ -1,5 +1,6 @@
-(ns athens.views.node-page
+(ns athens.views.pages.node-page
   (:require
+    ["@material-ui/core/Popover" :as Popover]
     ["@material-ui/icons/Bookmark" :default Bookmark]
     ["@material-ui/icons/BookmarkBorder" :default BookmarkBorder]
     ["@material-ui/icons/BubbleChart" :default BubbleChart]
@@ -9,23 +10,23 @@
     ["@material-ui/icons/Link" :default Link]
     ["@material-ui/icons/MoreHoriz" :default MoreHoriz]
     [athens.db :as db :refer [get-linked-references get-unlinked-references]]
-    [athens.keybindings :refer [destruct-key-down arrow-key-direction block-start? block-end?]]
     [athens.parse-renderer :as parse-renderer :refer [pull-node-from-string parse-and-render]]
     [athens.patterns :as patterns]
     [athens.router :refer [navigate-uid navigate]]
-    [athens.style :refer [color]]
+    [athens.style :refer [color DEPTH-SHADOWS]]
     [athens.util :refer [now-ts gen-block-uid escape-str is-daily-note get-caret-position recursively-modify-block-for-embed]]
     [athens.views.alerts :refer [alert-component]]
-    [athens.views.blocks :refer [block-el bullet-style]]
+    [athens.views.blocks.bullet :as bullet]
+    [athens.views.blocks.core :as blocks]
+    [athens.views.blocks.textarea-keydown :as textarea-keydown]
     [athens.views.breadcrumbs :refer [breadcrumbs-list breadcrumb]]
     [athens.views.buttons :refer [button]]
-    [athens.views.dropdown :refer [dropdown-style menu-style menu-separator-style]]
+    [athens.views.dropdown :refer [menu-style menu-separator-style]]
     [cljsjs.react]
     [cljsjs.react.dom]
     [clojure.string :as str]
     [datascript.core :as d]
     [garden.selectors :as selectors]
-    [goog.events :refer [listen unlisten]]
     [komponentit.autosize :as autosize]
     [re-frame.core :refer [dispatch subscribe]]
     [reagent.core :as r]
@@ -33,6 +34,12 @@
   (:import
     (goog.events
       KeyCodes)))
+
+;;-------------------------------------------------------------------
+;;--- material ui ---
+
+
+(def m-popover (r/adapt-react-class (.-default Popover)))
 
 ;;; Styles
 
@@ -42,6 +49,18 @@
    :padding "1rem 2rem 10rem 2rem"
    :flex-basis "100%"
    :max-width "55rem"})
+
+
+(def dropdown-style
+  {::stylefy/manual [[:.menu {:background (color :background-plus-2)
+                              :border-radius "calc(0.25rem + 0.25rem)" ;; Button corner radius + container padding makes "concentric" container radius
+                              :padding "0.25rem"
+                              :display "inline-flex"
+                              :box-shadow [[(:64 DEPTH-SHADOWS) ", 0 0 0 1px rgba(0, 0, 0, 0.05)"]]}]]})
+
+
+(def page-header-style
+  {:position "relative"})
 
 
 (def title-style
@@ -136,12 +155,11 @@
 
 (def page-menu-toggle-style
   {:position "absolute"
-   :left "-0.5rem"
+   :left "-1.5rem"
    :border-radius "1000px"
    :padding "0.375rem 0.5rem"
    :color (color :body-text-color :opacity-high)
-   :top "50%"
-   :transform "translate(-100%, -50%)"})
+   :top "0.5rem"})
 
 
 ;;; Helpers
@@ -165,7 +183,7 @@
   (.. e preventDefault)
   (let [node-page  (.. e -target (closest ".node-page"))
         block-page (.. e -target (closest ".block-page"))
-        {:keys [start value]} (destruct-key-down e)]
+        {:keys [start value]} (textarea-keydown/destruct-key-down e)]
     (cond
       block-page (dispatch [:split-block-to-children uid value start])
       node-page (if (empty? children)
@@ -175,9 +193,9 @@
 
 (defn handle-page-arrow-key
   [e uid state]
-  (let [{:keys [key-code target]} (destruct-key-down e)
-        start?          (block-start? e)
-        end?            (block-end? e)
+  (let [{:keys [key-code target]} (textarea-keydown/destruct-key-down e)
+        start?          (textarea-keydown/block-start? e)
+        end?            (textarea-keydown/block-end? e)
         {caret-position :caret-position} @state
         textarea-height (.. target -offsetHeight)
         {:keys [top height]} caret-position
@@ -201,11 +219,11 @@
 
 (defn handle-key-down
   [e uid state children]
-  (let [{:keys [key-code shift]} (destruct-key-down e)
+  (let [{:keys [key-code shift]} (textarea-keydown/destruct-key-down e)
         caret-position (get-caret-position (.. e -target))]
     (swap! state assoc :caret-position caret-position)
     (cond
-      (arrow-key-direction e) (handle-page-arrow-key e uid state)
+      (textarea-keydown/arrow-key-direction e) (handle-page-arrow-key e uid state)
       (and (not shift) (= key-code KeyCodes.ENTER)) (handle-enter e uid state children))))
 
 
@@ -312,7 +330,7 @@
   [parent-uid]
   [:div {:class "block-container"}
    [:div {:style {:display "flex"}}
-    [:span (use-style bullet-style)]
+    [:span (use-style bullet/bullet-style)]
     [:span {:on-click #(handle-new-first-child-block-click parent-uid)} "Click here to add content..."]]])
 
 
@@ -340,48 +358,49 @@
 
 
 (defn menu-dropdown
-  [_node state _daily-note?]
-  (let [ref                  (atom nil)
-        handle-click-outside (fn [e]
-                               (when (and (:menu/show @state)
-                                          (not (.. @ref (contains (.. e -target)))))
-                                 (swap! state assoc :menu/show false)))]
-    (r/create-class
-      {:display-name           "node-page-menu"
-       :component-did-mount    (fn [_this] (listen js/document "mousedown" handle-click-outside))
-       :component-will-unmount (fn [_this] (unlisten js/document "mousedown" handle-click-outside))
-       :reagent-render         (fn [node state daily-note?]
-                                 (let [{:block/keys [uid] sidebar :page/sidebar title :node/title} node
-                                       {:menu/keys [show]} @state]
-                                   (when show
-                                     [:div (merge (use-style dropdown-style
-                                                             {:ref #(reset! ref %)})
-                                                  {:style {:font-size "14px"
-                                                           :position  "absolute"
-                                                           :left      "-3em"
-                                                           :top       "3.5em"}})
-                                      [:div (use-style menu-style)
-                                       [:<>
-                                        (if sidebar
-                                          [button {:on-click #(dispatch [:page/remove-shortcut uid])}
-                                           [:<>
-                                            [:> BookmarkBorder]
-                                            [:span "Remove Shortcut"]]]
-                                          [button {:on-click #(dispatch [:page/add-shortcut uid])}
-                                           [:<>
-                                            [:> Bookmark]
-                                            [:span "Add Shortcut"]]])
-                                        [button {:on-click #(dispatch [:right-sidebar/open-item uid true])}
-                                         [:<>
-                                          [:> BubbleChart]
-                                          [:span "Show Local Graph"]]]]
-                                       [:hr (use-style menu-separator-style)]
-                                       [button {:on-click #(if daily-note?
-                                                             (dispatch [:daily-note/delete uid title])
-                                                             (do
-                                                               (navigate :pages)
-                                                               (dispatch [:page/delete uid title])))}
-                                        [:<> [:> Delete] [:span "Delete Page"]]]]])))})))
+  [node daily-note?]
+  (let [{:block/keys [uid] sidebar :page/sidebar title :node/title} node]
+    (r/with-let [ele (r/atom nil)]
+                [:<>
+                 [button {:class    [(when @ele "is-active")]
+                          :on-click #(reset! ele (.-currentTarget %))
+                          :style    page-menu-toggle-style}
+                  [:> MoreHoriz]]
+                 [m-popover
+                  (merge (use-style dropdown-style)
+                         {:style {:font-size "14px"}
+                          :open            @ele
+                          :anchorEl        @ele
+                          :onClose         #(reset! ele nil)
+                          :anchorOrigin    #js{:vertical   "bottom"
+                                               :horizontal "left"}
+                          :marginThreshold 10
+                          :transformOrigin #js{:vertical   "top"
+                                               :horizontal "left"}
+                          :classes {:root "backdrop"
+                                    :paper "menu"}})
+                  [:div (use-style menu-style)
+                   [:<>
+                    (if sidebar
+                      [button {:on-click #(dispatch [:page/remove-shortcut uid])}
+                       [:<>
+                        [:> BookmarkBorder]
+                        [:span "Remove Shortcut"]]]
+                      [button {:on-click #(dispatch [:page/add-shortcut uid])}
+                       [:<>
+                        [:> Bookmark]
+                        [:span "Add Shortcut"]]])
+                    [button {:on-click #(dispatch [:right-sidebar/open-item uid true])}
+                     [:<>
+                      [:> BubbleChart]
+                      [:span "Show Local Graph"]]]]
+                   [:hr (use-style menu-separator-style)]
+                   [button {:on-click #(if daily-note?
+                                         (dispatch [:daily-note/delete uid title])
+                                         (do
+                                           (navigate :pages)
+                                           (dispatch [:page/delete uid title])))}
+                    [:<> [:> Delete] [:span "Delete Page"]]]]]])))
 
 
 (defn ref-comp
@@ -406,7 +425,7 @@
                                             (swap! state assoc :block new-B :parents new-P)))}
                [parse-and-render (or title string) uid]]))]
          [:div.block-embed
-          [block-el
+          [blocks/block-el
            (recursively-modify-block-for-embed block embed-id)
            linked-ref-data
            {:block-embed? true}]]]))))
@@ -514,7 +533,7 @@
         unlinked-refs (r/atom [])]
     (fn [node editing-uid linked-refs]
       (let [{:block/keys [children uid] title :node/title} node
-            {:menu/keys [show] :alert/keys [message confirm-fn cancel-fn] alert-show :alert/show} @state
+            {:alert/keys [message confirm-fn cancel-fn] alert-show :alert/show} @state
             daily-note?  (is-daily-note uid)
             on-daily-notes? (= :home @(subscribe [:current-route/name]))]
 
@@ -530,44 +549,41 @@
                              :left     "35%"})
             [alert-component message confirm-fn cancel-fn]])
 
+
          ;; Header
-         [:h1 (use-style title-style
-                         {:data-uid uid
-                          :class    "page-header"
-                          :on-click (fn [e]
-                                      (.. e preventDefault)
-                                      (if (or daily-note? (.. e -shiftKey))
-                                        (navigate-uid uid e)
-                                        (dispatch [:editing/uid uid])))})
-          ;; Prevent editable textarea if a node/title is a date
-          ;; Don't allow title editing from daily notes, right sidebar, or node-page itself.
-          [button {:class    [(when show "active")]
-                   :on-click (fn [e]
-                               (.. e stopPropagation)
-                               (if show
-                                 (swap! state assoc :menu/show false)
-                                 (swap! state merge {:menu/show true})))
-                   :style    page-menu-toggle-style}
-           [:> MoreHoriz]]
-          (when-not daily-note?
-            [autosize/textarea
-             {:value       (:title/local @state)
-              :id          (str "editable-uid-" uid)
-              :class       (when (= editing-uid uid) "is-editing")
-              :on-blur     (fn [_]
-                             ;; add title Untitled-n for empty titles
-                             (when (empty? (:title/local @state))
-                               (swap! state assoc :title/local (auto-inc-untitled)))
-                             (handle-blur node state linked-refs))
-              :on-key-down (fn [e] (handle-key-down e uid state children))
-              :on-change   (fn [e] (handle-change e state))}])
-          ;; empty word break to keep span on full height else it will collapse to 0 height (weird ui)
-          (if (str/blank? (:title/local @state))
-            [:wbr]
-            [parse-renderer/parse-and-render (:title/local @state) uid])
+         [:header (use-style page-header-style)
 
           ;; Dropdown
-          [menu-dropdown node state daily-note?]]
+          [menu-dropdown node state daily-note?]
+
+          [:h1 (use-style title-style
+                          {:data-uid uid
+                           :class    "page-header"
+                           :on-click (fn [e]
+                                       (.. e preventDefault)
+                                       (if (or daily-note? (.. e -shiftKey))
+                                         (navigate-uid uid e)
+                                         (dispatch [:editing/uid uid])))})
+          ;; Prevent editable textarea if a node/title is a date
+          ;; Don't allow title editing from daily notes, right sidebar, or node-page itself.
+
+
+           (when-not daily-note?
+             [autosize/textarea
+              {:value       (:title/local @state)
+               :id          (str "editable-uid-" uid)
+               :class       (when (= editing-uid uid) "is-editing")
+               :on-blur     (fn [_]
+                             ;; add title Untitled-n for empty titles
+                              (when (empty? (:title/local @state))
+                                (swap! state assoc :title/local (auto-inc-untitled)))
+                              (handle-blur node state linked-refs))
+               :on-key-down (fn [e] (handle-key-down e uid state children))
+               :on-change   (fn [e] (handle-change e state))}])
+          ;; empty word break to keep span on full height else it will collapse to 0 height (weird ui)
+           (if (str/blank? (:title/local @state))
+             [:wbr]
+             [parse-renderer/parse-and-render (:title/local @state) uid])]]
 
          ;; Children
          (if (empty? children)
@@ -575,14 +591,14 @@
            [:div
             (for [{:block/keys [uid] :as child} children]
               ^{:key uid}
-              [block-el child])])
+              [blocks/block-el child])])
 
          ;; References
          [linked-ref-el state on-daily-notes? linked-refs]
          [unlinked-ref-el state on-daily-notes? unlinked-refs title]]))))
 
 
-(defn node-page-component
+(defn page
   [ident]
   (let [{:keys [#_block/uid node/title] :as node} (db/get-node-document ident)
         editing-uid   @(subscribe [:editing/uid])